requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
anthropic>=0.7.0 